+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+# <a href='https://github.com/mrjackwills/mealpedant_vue/releases/tag/v2.2.0'>v2.2.0</a>
+### 2026-08-27
+
+### Chores
++ _typos.toml update, [cf559370](https://github.com/mrjackwills/mealpedant_vue/commit/cf5593707c532917719354b9fbd3abf4354d2161)
++ .devcontainer updated, [1260567d](https://github.com/mrjackwills/mealpedant_vue/commit/1260567d47bc880b6de8f969b5704903c6a80a1e)
++ dependencies updated, [375da7a0](https://github.com/mrjackwills/mealpedant_vue/commit/375da7a08d1fe0989c163c0caac9631583b71b04), [9058dbf8](https://github.com/mrjackwills/mealpedant_vue/commit/9058dbf8ce944c24c58dfdf60ddd5544efd137dc)
++ GitHub workflow updated, [7d1a4f05](https://github.com/mrjackwills/mealpedant_vue/commit/7d1a4f054d5a7a8ff77f937f09b28706783ac67c)
+
+### Features
++ percentages() in helpers, [0fdef589](https://github.com/mrjackwills/mealpedant_vue/commit/0fdef589b3eef4a3414ed4cd19fe99b1d0c158c6)
++ set pwa status, [f6cddf3f](https://github.com/mrjackwills/mealpedant_vue/commit/f6cddf3f96f925e4b50a86fd34076e49e7697812)
++ unselectable, [3cc0f0a7](https://github.com/mrjackwills/mealpedant_vue/commit/3cc0f0a7e3a964b2cf0d295758423c99097d1176)
++ Vuetify prop typos, [2c73568c](https://github.com/mrjackwills/mealpedant_vue/commit/2c73568c8570386243cab6d1c7c7b28e668196a6), [82800a37](https://github.com/mrjackwills/mealpedant_vue/commit/82800a37af3ea2d5db94227ac5c18c3365e4569b)
++ require password when enableding 2FA always_required, [6183e39c](https://github.com/mrjackwills/mealpedant_vue/commit/6183e39c2284ebdf058a0f8cb403a5fc886380b9)
+
+### Fixes
++ AdminBackup typo, [17bde3da](https://github.com/mrjackwills/mealpedant_vue/commit/17bde3da9a182cdd8c30f6b5b5630f00a0c4f79f)
++ AdminView missing await, [ec498ead](https://github.com/mrjackwills/mealpedant_vue/commit/ec498ead76bbbb4bf6e7e8caa32cc62e5f680ced)
++ CustomSearchSelector, add debounce to search, [d1275875](https://github.com/mrjackwills/mealpedant_vue/commit/d1275875ed447495c868794e0f2b06f5b292e4b6)
++ fetch.ts mealhash_get @wrap, [d77c2591](https://github.com/mrjackwills/mealpedant_vue/commit/d77c25914ad62a11297c1a10e293d1246300564f)
++ ForgotPassword email validation, [d0b5c2b8](https://github.com/mrjackwills/mealpedant_vue/commit/d0b5c2b808c0b0cb2a6c04fbc23108e6d386f54c)
++ helpers.rs second plural, [730257db](https://github.com/mrjackwills/mealpedant_vue/commit/730257db205e7d3ccfe93b83424d6d933a6659a6)
++ missing TPerson use in AdminPhotos, [dd5d4eeb](https://github.com/mrjackwills/mealpedant_vue/commit/dd5d4eeb29946d50f86fcb65152f354af536551b)
++ password length errormessages, [11a3544b](https://github.com/mrjackwills/mealpedant_vue/commit/11a3544b7917442d9a31da3dee5e20c08fba3d4c)
++ require auth for 2fa backups, [14b095d0](https://github.com/mrjackwills/mealpedant_vue/commit/14b095d014dffa0d0bba11cc4fce7cc1f5671770)
++ ResetView validations, [099949a7](https://github.com/mrjackwills/mealpedant_vue/commit/099949a70ace66d903ad4093952d6d070e3244fd)
++ router hexRegister, [57ee098a](https://github.com/mrjackwills/mealpedant_vue/commit/57ee098a4737dbcb72550d899bb8a12663e014e0)
++ router typo, [3f095239](https://github.com/mrjackwills/mealpedant_vue/commit/3f0952399f0bf5c3829321792b42c1dd9412108c)
++ SingleMeal missing await, [5af0753f](https://github.com/mrjackwills/mealpedant_vue/commit/5af0753f049bbb7572721b8371a30be9af0f9b32)
++ SingleMeal pointless :col, [4350f3d7](https://github.com/mrjackwills/mealpedant_vue/commit/4350f3d74b7e0d1f9d86402613ca5520209090c4)
++ SingleMeal.vue fileInserted(), [c392c61d](https://github.com/mrjackwills/mealpedant_vue/commit/c392c61d830de34235c8fe793ee919013d95a267)
++ small store bugs, [7cd19f16](https://github.com/mrjackwills/mealpedant_vue/commit/7cd19f1668bb19ec72d4830335eeea5236233206)
++ start & end dates use mealSearch data, [c402d8ca](https://github.com/mrjackwills/mealpedant_vue/commit/c402d8caca2aacb6225e23bf79128a5c78949e43)
++ store/meal uncompress_search_by, [36d220de](https://github.com/mrjackwills/mealpedant_vue/commit/36d220dea9762ba1a2585f9c15d1c377e24dbc9e), [4eab59a7](https://github.com/mrjackwills/mealpedant_vue/commit/4eab59a702f27c4917aa78182f99d2efc0f057b8)
++ typos, [0359499e](https://github.com/mrjackwills/mealpedant_vue/commit/0359499e5053f11a2c430e1294123058954cfa99)
++ useHeight(), [9f2b2d41](https://github.com/mrjackwills/mealpedant_vue/commit/9f2b2d41fef4e573401c742e2c94b75b098fcc2d)
+
+### Refactors
++ dead code removed, [6d806cea](https://github.com/mrjackwills/mealpedant_vue/commit/6d806cea741c36c92f7cafde64a6904174609a2d), [db85eeac](https://github.com/mrjackwills/mealpedant_vue/commit/db85eeaccf8bdda14d57e71b8eca22f75e7cf4b7)
++ meal search, [b662cf19](https://github.com/mrjackwills/mealpedant_vue/commit/b662cf19c25b507a827a74340f0212d7f62b92f6)
++ mealStorage.ts, [958f31e8](https://github.com/mrjackwills/mealpedant_vue/commit/958f31e8728bb7716a7bc26b22254375f8a0de8f)
++ registerLinks put in globalConsts, [49c4d92e](https://github.com/mrjackwills/mealpedant_vue/commit/49c4d92e90aa11c1b995f58f8597c3347d53f473)
+
 # <a href='https://github.com/mrjackwills/mealpedant_vue/releases/tag/v2.1.4'>v2.1.4</a>
 ### 2026-07-16